dms-backend/requirements.txt
cjm_pil_utils==0.0.9
cjm_pytorch_utils==0.0.6
django==5.0.4
djangorestframework==3.15.1
Faker==24.9.0
matplotlib==3.8.4
numpy==1.26.4
onnxruntime==1.17.3
pandas==2.2.2
Pillow==10.3.0
scipy==1.13.0
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
